pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-triangle-reader"
version = "0.1.0"
description = "Simple PDF word-by-word reader with a triangle pointer."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyMuPDF>=1.24,<2.0",
  "Pillow>=10.0,<12.0",
]

[project.scripts]
pdf-triangle-reader = "pdf_triangle_reader.app:main"

[tool.setuptools]
packages = ["pdf_triangle_reader"]
